Factor out key parts of stager to bin/bundle_tools
